call set -u only after env vars are tested
This commit fixes a race condition where we search for variable
BASH_SOURCE, but it is not set. In that case the update.sh would abort.
But it should not abort, because in the following lines we are going to
set DIR variable by another way.
To avoid this case where BASH_SOURCE is not set, the `set -u` command is
moved after the code, where we expect all variables to be valid.