Make RSS feed entry links valid and accessible
According to the ATOM spec link elements must have a href:
https://validator.w3.org/feed/docs/atom.html

Also most feed readers (TinyTiny, Feedly, Thunderbird) prefer link elements in
this order:

- plain
- with attribute rel=alternate
- with attribute rel=via

As the content is work of the entry author, link to it by default. Not to the
wallabag instance entry.

This also makes it possible to share links with people who don't have access to
the wallabag instance via the RSS feed.