fix: inappropriate handling of escaped query arguments
The list(s) endpoints receive a JSON encoded list of URLs as a query
param. Because it's a query param, it's escaped, so the `json_decode`
method can't decode it without first calling `urldecode`.
Also, the `POST` versions were accessing the elments as if they were the
strings, rather than objects of the shape `{'url':'<url>'}`.